+++ src/Armature.Core/src/BuildChainPatterns/IfFirstUnit.cs
@@ -14,7 +14,7 @@ public class IfFirstUnit : BuildChainPatternByUnitBase
   /// Checks if the first unit in the build chain matches the specified patter.
   /// If it is the target unit, returns build actions for it, if no, pass the rest of the build chain to each child and returns all actions from children merged
   /// </summary>
-  public override bool GatherBuildActions(BuildChain buildChain, out WeightedBuildActionBag? actionBag, int inputWeight)
+  public override bool GatherBuildActions(BuildChain buildChain, out WeightedBuildActionBag? actionBag, long inputWeight)
   {
     actionBag = null;
