+++ src/Armature.Core/src/BuildChainPatterns/SkipWhileUnit.cs
@@ -10,7 +10,7 @@ public class SkipWhileUnit : BuildChainPatternByUnitBase
   public SkipWhileUnit(IUnitPattern pattern) : base(pattern, WeightOf.BuildChainPattern.SkipWhileUnit) { }
   public SkipWhileUnit(IUnitPattern unitPattern, int weight) : base(unitPattern, weight) { }
 
-  public override bool GatherBuildActions(BuildChain buildChain, out WeightedBuildActionBag? actionBag, int inputWeight)
+  public override bool GatherBuildActions(BuildChain buildChain, out WeightedBuildActionBag? actionBag, long inputWeight)
   {
     var hasActions = false;
